♻️ Refactor :  파일 책임 분리
