+++ src/navigation/stacks/HomeStackNavigator.tsx
@@ -5,7 +5,7 @@ import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import HomeTabScreens from '../tabs/TabNavigator';
 import { HomeNavigationProps } from '../types/navigationTypes';
 
-import EditSchedule from '@/features/calendar/editSchedule/components/EditSchedule';
+import CalendarScheduleScreen from '@/screens/Calendar/CalendarSchedule';
 import HotBoardScreen from '@/screens/Home/HotBoard';
 import HomeSearchScreen from '@/screens/Home/Search';
 
@@ -23,9 +23,8 @@ const HomeRoute = () => {
       <Stack.Screen name="Detail" component={DetailScreen} />
       <Stack.Screen name="Profile" component={ProfileScreen} />
 
-      {/* Calendar Stack 내의 EditSchedule 및 CreateSchedule 화면 추가 */}
-      <Stack.Screen name="EditSchedule" component={EditSchedule} />
-      <Stack.Screen name="CreateSchedule" component={EditSchedule} />
+      <Stack.Screen name="EditSchedule" component={CalendarScheduleScreen} />
+      <Stack.Screen name="CreateSchedule" component={CalendarScheduleScreen} />
     </Stack.Navigator>
   );
 };
